05_engage_nlg_rag/requirements.txt
openai==0.28.1
streamlit
streamlit-option-menu
streamlit-extras
faiss-cpu
langchain
langchain-community
scipy
scikit-learn
numpy